requirements.txt
fastapi==0.111.0
langchain==0.2.6
langchain_community==0.2.6
langchain_experimental==0.0.62
langchain_groq==0.1.6
langchain_openai==0.1.13
matplotlib==3.8.4
numpy==2.0.0
pandas==2.2.2
Pillow==10.3.0
prophet==1.1.5
pydantic==2.7.4
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.5.0
scipy==1.14.0
SQLAlchemy==2.0.30
streamlit==1.36.0
streamlit_shadcn_ui==0.1.18
uvicorn==0.30.1
